TDW-1646 move admiral to dependencies
