+++ scripts/scripts.js
@@ -1053,11 +1053,21 @@ async function loadEager(doc) {
 }
 
 /**
+ * Loads everything that required HTML to be in place.
+ */
+function loadDependencies() {
+  // eslint-disable-next-line import/no-cycle
+  import('./dependencies.js');
+}
+
+/**
  * loads everything that doesn't need to be delayed.
  */
 async function loadLazy(doc) {
   const main = doc.querySelector('main');
   await loadBlocks(main);
+
+  loadDependencies();
 
   const { hash } = window.location;
   const element = hash ? main.querySelector(hash) : false;
